meeseeksai/pyproject.toml
[tool.poetry]
name = "meeseeksai"
version = "0.1.6"
description = ""
authors = ["Jaya <s.k.jayachithra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "*"
python-dotenv = "^1.0.0"
langchain = "^0.0.281"
tiktoken = "^0.4.0"

[tool.poetry.scripts]
meeseeks = "meeseeksai.meeseeksai:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
